backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aws-lambda-powertools = "^3.4.1"
boto3 = "^1.35.99"

[tool.poetry.group.dev.dependencies]
aws-cdk-lib = "^2.175.1"
constructs = "^10.4.2"
pytest = "^8.3.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
